+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+### Added
+- Accessing a non-empty multi-selection field in a template now returns the new `CategoryGroupCollection`, which is designed to mimic a Craft element query (though doesn't yet support the more advanced filtering options that element queries have) (for now, iterating directly over a `CategoryGroupCollection` in templates is still supported, so no template changes are needed at this stage, but it will be deprecated in the future)
+- Added `all()`, `one()`, `nth()`, `count()` and `ids()` methods for `CategoryGroupCollection`
+
 ## 1.1.0 - 2019-09-29
 ### Added
 - Added the `singleSelectionDefault` config setting